+++ src/qnethernet/security/entropy.cpp
@@ -21,7 +21,7 @@
 #include "qnethernet/internal/macro_funcs.h"
 #include "qnethernet/platforms/pgmspace.h"
 
-enum TRNGValues {
+enum TRNGValues : uint32_t {
   // Clock settings
   TRNG_CONFIG_CLOCK_MODE   = 0,  // 0=Ring Oscillator, 1=System Clock (test use only)
   TRNG_CONFIG_RING_OSC_DIV =     // Divide by 2^n
@@ -231,7 +231,7 @@ static bool fillEntropy() {
 
 // // Reads a single entropy byte.
 // ATTRIBUTE_NODISCARD
-// static bool readEntropy(uint8_t* b) {
+// static bool readEntropy(uint8_t* const b) {
 //   if (!fillEntropy()) {
 //     return false;
 //   }
@@ -303,28 +303,30 @@ static inline bool random32(uint32_t* const r) {
 
 extern "C" {
 
-uint32_t entropy_random() {
+bool entropy_random(uint32_t* const out) {
   uint32_t r;
   if (!random32(&r)) {
-    return 0;
+    return false;
   }
-  return r;
+  *out = r;
+  return true;
 }
 
-uint32_t entropy_random_range(const uint32_t range) {
+bool entropy_random_range(const uint32_t range, uint32_t* const out) {
   if (range == 0) {
     errno = EDOM;
-    return 0;
+    return false;
   }
 
   uint32_t r;
   if (!random32(&r)) {
-    return 0;
+    return false;
   }
 
   // Is power of 2?
   if ((range & (range - 1)) == 0) {
-    return r & (range - 1);
+    *out = r & (range - 1);
+    return true;
   }
 
   // Daniel Lemire's nearly-divisionless algorithm
@@ -337,13 +339,14 @@ uint32_t entropy_random_range(const uint32_t range) {
     const uint32_t threshold = -range % range;  // 2^L mod s = (2^L − s) mod s
     while (low < threshold) {
       if (!random32(&r)) {
-        return 0;
+        return false;
       }
       product = uint64_t{r} * uint64_t{range};
       low = static_cast<uint32_t>(product);
     }
   }
-  return static_cast<uint32_t>(product >> 32);
+  *out = static_cast<uint32_t>(product >> 32);
+  return true;
 }
 
 }  // extern "C"