+++ src/qnethernet/security/entropy.h
@@ -42,18 +42,22 @@ size_t trng_available(void);
 // this will return a value less than 'size'.
 size_t trng_data(void* data, size_t size);
 
-// Returns a random 4-byte number from the entropy generator. If there was any
-// entropy generation error then errno will be set to EIO.
-uint32_t entropy_random(void);
+// Generates a random 4-byte number and returns whether successful. The result
+// is stored in '*out'. If there was any entropy generation error then errno
+// will be set to EIO and this returns false.
+bool entropy_random(uint32_t* out);
 
-// Returns a random number in the range [0, range). This uses an unbiased
-// algorithm. If there was any entropy generation error then errno will be set
-// to EIO and this will return zero. If 'range' is zero then errno will be set
-// to EDOM.
+// Generates a random number in the range [0, range) and returns whether
+// successful. This uses an unbiased algorithm. The result is stored in '*out'.
 //
-// See: http://www.adammil.net/blog/v134_Efficiently_generating_random_numbers_without_bias.html
-// See: https://lemire.me/blog/2019/09/28/doubling-the-speed-of-stduniform_int_distribution-in-the-gnu-c-library/
-uint32_t entropy_random_range(uint32_t range);
+// This returns false if:
+// * There was any entropy generation error &mdash; errno will be set to EIO
+// * 'range' is zero &mdash; errno will be set to EDOM
+//
+// See:
+// * http://www.adammil.net/blog/v134_Efficiently_generating_random_numbers_without_bias.html
+// * https://lemire.me/blog/2019/09/28/doubling-the-speed-of-stduniform_int_distribution-in-the-gnu-c-library/
+bool entropy_random_range(uint32_t range, uint32_t* out);
 
 #ifdef __cplusplus
 }  // extern "C"