+++ src/qnethernet/security/random_device.cpp
@@ -7,7 +7,7 @@
 #include "qnethernet/security/random_device.h"
 
 // C++ includes
-#include <cstdint>
+#include <stdexcept>
 
 #include "qnethernet/platforms/pgmspace.h"
 
@@ -17,7 +17,7 @@ namespace security {
 extern "C" {
 void qnethernet_hal_init_entropy();
 void qnethernet_hal_deinit_entropy();
-double qnethernet_hal_estimate_entropy();
+double qnethernet_hal_estimate_entropy(size_t typeSize);
 size_t qnethernet_hal_entropy_available();
 uint32_t qnethernet_hal_entropy();
 size_t qnethernet_hal_fill_entropy(void* buf, size_t size);
@@ -40,15 +40,23 @@ random_device::random_device(const std::string& token) {
 }
 
 double random_device::entropy() const noexcept {
-  return qnethernet_hal_estimate_entropy();
+  return qnethernet_hal_estimate_entropy(sizeof(result_type));
 }
 
 random_device::result_type random_device::operator()() {
-  return static_cast<result_type>(qnethernet_hal_entropy());
+  result_type r;
+  (*this)(&r, sizeof(r));
+  return r;
 }
 
-size_t random_device::operator()(uint8_t* const buf, const size_t size) {
-  return qnethernet_hal_fill_entropy(buf, size);
+void random_device::operator()(void* const buf, const size_t size) {
+  if (qnethernet_hal_fill_entropy(buf, size) != size) {
+#if defined(__cpp_exceptions)
+    throw std::runtime_error("generation");
+#else
+    std::__throw_runtime_error("generation");
+#endif  // __cpp_exceptions
+  }
 }
 
 size_t random_device::available() {